Decompose data-input.js into smaller files.
